Organized files into src, include, and test folders. Also added a test (that has syntax errors?)
